+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelogs
 
+### 0.3.3
+
+-   Run prettier on created components
+
 ### 0.3.2
 
-- Add tests to ensure more stability
+-   Add tests to ensure more stability
 
 ### 0.3.0